test(representation): réaligne la couverture après la centralisation des helpers
