+++ packages/app/src/modules/declaration-representation/steps/__tests__/Step5Review.restitution.test.tsx
@@ -1,37 +1,12 @@
-import { act, render, screen, within } from "@testing-library/react";
-import userEvent from "@testing-library/user-event";
-import { beforeEach, describe, expect, it, vi } from "vitest";
-
-const { push, submitMutate, openModal, closeModal, submitState } = vi.hoisted(
-	() => ({
-		push: vi.fn(),
-		submitMutate: vi.fn(),
-		openModal: vi.fn(),
-		closeModal: vi.fn(),
-		submitState: {
-			error: null as { message: string } | null,
-			isPending: false,
-			onSuccess: undefined as (() => void) | undefined,
-		},
-	}),
-);
-
-vi.mock("next/navigation", () => ({
-	usePathname: vi.fn(),
-	useRouter: () => ({
-		push,
-		replace: vi.fn(),
-		back: vi.fn(),
-		refresh: vi.fn(),
-	}),
-}));
+import { screen, within } from "@testing-library/react";
+import { describe, expect, it, vi } from "vitest";
 
 vi.mock("~/modules/shared", async (importOriginal) => ({
 	...(await importOriginal<typeof import("~/modules/shared")>()),
 	useDsfrModal: () => ({
 		modalRef: { current: null },
-		open: openModal,
-		close: closeModal,
+		open: vi.fn(),
+		close: vi.fn(),
 	}),
 }));
 
@@ -39,84 +14,39 @@ vi.mock("~/trpc/react", () => ({
 	api: {
 		representationDeclaration: {
 			submit: {
-				useMutation: ({ onSuccess }: { onSuccess: () => void }) => {
-					submitState.onSuccess = onSuccess;
-					return {
-						error: submitState.error,
-						isPending: submitState.isPending,
-						mutate: submitMutate,
-					};
-				},
+				useMutation: () => ({
+					error: null,
+					isPending: false,
+					mutate: vi.fn(),
+				}),
 			},
 		},
 	},
 }));
 
 import {
-	COMPUTABLE_EXECUTIVES,
 	COMPUTABLE_MEMBERS,
 	NO_EXECUTIVES,
-	NO_MANAGEMENT_BODY,
 	NON_COMPLIANT_EXECUTIVES,
 	NON_COMPLIANT_MEMBERS,
 	OFFLINE_PUBLICATION,
 	REPRESENTATION_YEAR,
 	SINGLE_EXECUTIVE,
-	VALID_REFERENCE_PERIOD,
 	WEBSITE_PUBLICATION,
 } from "~/modules/declaration-representation/__tests__/fixtures";
-import type { RepresentationDraftContextValue } from "~/modules/declaration-representation/shared/draft/DraftContext";
-import { RepresentationDraftProvider } from "~/modules/declaration-representation/shared/draft/DraftContext";
 import {
 	EXECUTIVES_TITLE,
 	MEMBERS_TITLE,
 } from "~/modules/declaration-representation/shared/reviewSummary";
-import type { RepresentationDraft } from "~/modules/declaration-representation/types";
-import { TOTAL_REPRESENTATION_STEPS } from "~/modules/declaration-representation/types";
-import { Step5Review } from "../Step5Review";
+import {
+	BOTH_COMPUTABLE,
+	FULLY_COMPLIANT,
+	NOTHING_COMPUTABLE,
+	renderReview,
+	SINGLE_GAP,
+} from "./step5ReviewHarness";
 
-const PREVIOUS_HREF = "/declaration-representation/etape/4";
-const CONFIRMATION_HREF = "/declaration-representation/confirmation";
 const TELEACCORDS_URL = "https://www.teleaccords.travail.gouv.fr";
-
-const BOTH_COMPUTABLE = { ...COMPUTABLE_EXECUTIVES, ...COMPUTABLE_MEMBERS };
-const NOTHING_COMPUTABLE = { ...NO_EXECUTIVES, ...NO_MANAGEMENT_BODY };
-const FULLY_COMPLIANT = { ...BOTH_COMPUTABLE, ...WEBSITE_PUBLICATION };
-const SINGLE_GAP = {
-	...COMPUTABLE_EXECUTIVES,
-	...NON_COMPLIANT_MEMBERS,
-	...WEBSITE_PUBLICATION,
-};
-
-function renderReview({
-	draft = {},
-	isReadOnly = false,
-}: {
-	draft?: Partial<RepresentationDraft>;
-	isReadOnly?: boolean;
-} = {}) {
-	const value: RepresentationDraftContextValue = {
-		year: REPRESENTATION_YEAR,
-		step: TOTAL_REPRESENTATION_STEPS,
-		draft: {
-			currentStep: TOTAL_REPRESENTATION_STEPS,
-			...VALID_REFERENCE_PERIOD,
-			...draft,
-		},
-		setDraftValues: vi.fn(),
-		isSaving: false,
-		isPendingSave: false,
-		isReadOnly,
-		previousHref: PREVIOUS_HREF,
-		registerStepValidator: vi.fn(),
-	};
-
-	return render(
-		<RepresentationDraftProvider value={value}>
-			<Step5Review />
-		</RepresentationDraftProvider>,
-	);
-}
 
 function definitionValue(term: string): string {
 	const value = screen.getByText(term, { selector: "dt" }).nextElementSibling;
@@ -145,42 +75,6 @@ function complianceBadges() {
 		selector: "p.fr-badge",
 	});
 }
-
-function submitButton() {
-	return screen.queryByRole("button", { name: "Soumettre" });
-}
-
-// The DSFR modal is driven by the DSFR runtime, absent in jsdom: its <dialog>
-// stays closed, so its content has to be queried outside the a11y tree.
-function modalCheckbox() {
-	return screen.queryByRole("checkbox", { hidden: true });
-}
-
-function modalButton(name: string) {
-	return screen.getByRole("button", { hidden: true, name });
-}
-
-async function certify() {
-	const checkbox = modalCheckbox();
-	if (checkbox === null) throw new Error("Missing certification checkbox.");
-	await userEvent.click(checkbox);
-}
-
-async function openSubmitModal() {
-	const button = submitButton();
-	if (button === null) throw new Error("Missing submit button.");
-	await userEvent.click(button);
-}
-
-beforeEach(() => {
-	push.mockReset();
-	submitMutate.mockReset();
-	openModal.mockReset();
-	closeModal.mockReset();
-	submitState.error = null;
-	submitState.isPending = false;
-	submitState.onSuccess = undefined;
-});
 
 describe("Step5Review — restitution du brouillon (S18)", () => {
 	it("restitutes the reference year and period", () => {
@@ -311,7 +205,7 @@ describe("Step5Review — écarts non calculables (S12)", () => {
 		).not.toBeInTheDocument();
 	});
 
-	it("tells the modal that nothing is computable", async () => {
+	it("tells the modal that nothing is computable", () => {
 		renderReview({ draft: NOTHING_COMPUTABLE });
 
 		expect(
@@ -408,104 +302,5 @@ describe("Step5Review — prochaines étapes (S17)", () => {
 		expect(
 			screen.queryByRole("heading", { name: "Prochaines étapes" }),
 		).not.toBeInTheDocument();
-	});
-});
-
-describe("Step5Review — soumission (S19)", () => {
-	it("opens the confirmation modal from the CTA without submitting", async () => {
-		renderReview({ draft: FULLY_COMPLIANT });
-
-		await openSubmitModal();
-
-		expect(openModal).toHaveBeenCalledTimes(1);
-		expect(submitMutate).not.toHaveBeenCalled();
-	});
-
-	it("submits the draft payload once the declarant certifies", async () => {
-		renderReview({ draft: FULLY_COMPLIANT });
-
-		await openSubmitModal();
-		await certify();
-		await userEvent.click(modalButton("Valider"));
-
-		expect(submitMutate).toHaveBeenCalledWith({
-			payload: {
-				...VALID_REFERENCE_PERIOD,
-				...COMPUTABLE_EXECUTIVES,
-				...COMPUTABLE_MEMBERS,
-				...WEBSITE_PUBLICATION,
-			},
-			year: REPRESENTATION_YEAR,
-		});
-	});
-
-	it("sends no publication key when no gap is computable (S12)", async () => {
-		renderReview({ draft: { ...NOTHING_COMPUTABLE, ...WEBSITE_PUBLICATION } });
-
-		await openSubmitModal();
-		await certify();
-		await userEvent.click(modalButton("Valider"));
-
-		expect(submitMutate).toHaveBeenCalledWith({
-			payload: { ...VALID_REFERENCE_PERIOD, ...NOTHING_COMPUTABLE },
-			year: REPRESENTATION_YEAR,
-		});
-	});
-
-	it("keeps the submission locked until the declarant certifies", async () => {
-		renderReview({ draft: FULLY_COMPLIANT });
-
-		await openSubmitModal();
-
-		expect(modalButton("Valider")).toBeDisabled();
-	});
-
-	it("closes the modal and lands on the confirmation screen on success", () => {
-		renderReview({ draft: FULLY_COMPLIANT });
-
-		act(() => submitState.onSuccess?.());
-
-		expect(closeModal).toHaveBeenCalledTimes(1);
-		expect(push).toHaveBeenCalledWith(CONFIRMATION_HREF);
-	});
-
-	it("holds the modal while the submission is in flight", async () => {
-		submitState.isPending = true;
-		renderReview({ draft: FULLY_COMPLIANT });
-
-		await openSubmitModal();
-		await certify();
-
-		expect(modalButton("Envoi en cours…")).toBeDisabled();
-		expect(submitMutate).not.toHaveBeenCalled();
-	});
-
-	it("surfaces the submission failure without leaving the screen", () => {
-		submitState.error = { message: "La soumission a échoué." };
-		renderReview({ draft: FULLY_COMPLIANT });
-
-		expect(screen.getByRole("alert")).toHaveTextContent(
-			"La soumission a échoué.",
-		);
-		expect(push).not.toHaveBeenCalled();
-	});
-});
-
-describe("Step5Review — navigation", () => {
-	it("links back to the step the funnel came from", () => {
-		renderReview({ draft: FULLY_COMPLIANT });
-
-		expect(screen.getByRole("link", { name: "Précédent" })).toHaveAttribute(
-			"href",
-			PREVIOUS_HREF,
-		);
-	});
-
-	it("offers no submission while the campaign is closed (S23)", () => {
-		renderReview({ draft: FULLY_COMPLIANT, isReadOnly: true });
-
-		expect(submitButton()).not.toBeInTheDocument();
-		expect(modalCheckbox()).not.toBeInTheDocument();
-		expect(screen.getByRole("link", { name: "Précédent" })).toBeInTheDocument();
 	});
 });
